+++ src/Entity/BearerUser.php
@@ -1,12 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Paysera\BearerAuthenticationBundle\Entity;
 
 class BearerUser implements BearerUserInterface
 {
-    private $username;
-    private $token;
-    private $roles;
+    private string $username;
+    private string $token;
+    private array $roles;
 
     public function __construct($username, $token, array $roles = [])
     {
@@ -15,33 +17,41 @@ class BearerUser implements BearerUserInterface
         $this->roles = $roles;
     }
 
-    public function getRoles()
+    public function getRoles(): array
     {
         return $this->roles;
     }
 
-    public function getPassword()
+    public function getPassword(): ?string
     {
         // bearer has token
+        return null;
     }
 
-    public function getSalt()
+    public function getSalt(): ?string
     {
         // no salt
+
+        return null;
     }
 
-    public function getUsername()
+    public function getUsername(): string
     {
         return $this->username;
     }
 
-    public function eraseCredentials()
+    public function eraseCredentials(): void
     {
 
     }
 
-    public function getToken()
+    public function getToken(): string
     {
         return $this->token;
+    }
+
+    public function getUserIdentifier(): string
+    {
+        return $this->username;
     }
 }