+++ src/Entity/BearerUserInterface.php
@@ -1,13 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Paysera\BearerAuthenticationBundle\Entity;
 
 use Symfony\Component\Security\Core\User\UserInterface;
 
 interface BearerUserInterface extends UserInterface
 {
-    /**
-     * @return string
-     */
-    public function getToken();
+    public function getToken(): string;
 }