+++ src/Security/Authentication/Token/BearerTokenInterface.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Paysera\BearerAuthenticationBundle\Security\Authentication\Token;
 
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
@@ -9,5 +11,5 @@ interface BearerTokenInterface extends TokenInterface
     /**
      * @return string
      */
-    public function getToken();
+    public function getToken(): string;
 }