+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 2024-02-08
+### Added 
+- Added support for Symfony ^5.4
+- Dropped support for PHP 5.5 and Symfony 2.x and 3.x
+- Renamed Paysera\BearerAuthenticationBundle\DependencyInjection\Security\Factory\BearerFactory to Paysera\BearerAuthenticationBundle\DependencyInjection\Security\Factory\BearerSecurityFactory
+- Added Paysera\BearerAuthenticationBundle\DependencyInjection\Security\Factory\BearerAuthenticatorFactory for Symfony 5
+- Changed src/Security/Authentication/Token/BearerTokenInterface::getToken(), added return type
+
 ## 1.0.2
 ### Added
 - Deprecated \Symfony\Component\DependencyInjection\DefinitionDecorator class removed and